Fixed the bug in the collector for count queries & corrected the function used for getting count from the bucketset

+++ gemfire-connector/connector/src/main/scala/io/snappydata/spark/gemfire/connector/GemFireConnection.scala
@@ -66,7 +66,7 @@ trait GemFireConnection {
     */
   def getQuery(queryString: String): Query
 
-  def getCount(regionPath: String, buckets: Set[Int], whereClause: Option[String]): Int
+  def getCount(regionPath: String, buckets: Set[Int], whereClause: Option[String]): Long
 
   /** Close the connection */
   def close(): Unit
